requirements.txt
flask
flask-cors
flasgger
pyjwt
requests
factur-x
pypdf
cryptography
